fix: protect views

+++ announcement/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 from .models import Announcement
+from main.models import Store
 from django.utils.html import strip_tags
 from django.http import HttpResponse, HttpResponseRedirect
 from django.core import serializers
@@ -15,6 +16,11 @@ def create_announcement_ajax(request):
   title = strip_tags(request.POST.get("title"))
   description = strip_tags(request.POST.get("description"))
   store_id = request.POST.get("store_id")
+
+  user = request.user.ajeg_user
+  store = Store.objects.get(pk=store_id)
+  if store not in user.merchant_store:
+    return HttpResponse(b"UNAUTHORIZED", status=403)
 
   announcement = Announcement(title=title, description=description, store_id=store_id)
   announcement.save() 
@@ -44,5 +50,10 @@ def edit_announcement(request, id):
 
 def delete_announcement(request, id):
   announcement = Announcement.objects.get(pk=id)
+
+  user = request.user.ajeg_user
+  if announcement.store not in user.merchant_store:
+    return HttpResponse(b"UNAUTHORIZED", status=403)
+
   announcement.delete()
   return HttpResponseRedirect(reverse('main:show_main'))
